Lab7

+++ Lab5/5a/5a.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-    int a, b, c, x;
+    double a, b, c, x;
     printf("Введите значение a > ");
     scanf_s("%d", &a);
     printf("Введите значение b > ");
